refactor(cloud): single code block component that tracks copy events
